Renamed cmb_lensing.py to cmb.py
