+++ cloelib/summary_statistics/angular_two_point.py
@@ -4,7 +4,7 @@
 from cloelib.observables.tracer import Tracer
 from cloelib.observables.photo import PositionsTracer
 from cloelib.observables.photo import ShearTracer
-from cloelib.observables.cmb_lensing import CMBLensingTracer
+from cloelib.observables.cmb import CMBLensingTracer
 from cloelib.auxiliary.units import SPEED_OF_LIGHT
 from cloelib.auxiliary.math_utils import simpsons_weights_jit
 from cloelib.profiling import profile_function
